+++ APIBlox.NetCore.EventStore/EventStoreService.cs
@@ -20,8 +20,7 @@ namespace APIBlox.NetCore
         }
 
         public async Task<EventStreamModel> WriteToEventStreamAsync(string streamId, EventModel[] events,
-            long? expectedVersion = null, object metadata = null,
-            CancellationToken cancellationToken = default
+            long? expectedVersion = null, CancellationToken cancellationToken = default
         )
         {
             if (streamId.IsEmptyNullOrWhiteSpace())